clean up types and dedupe rowKeys
